chore: rename passkey migration and group passkey files
Migration 0002_material_scorpion to 0002_passkey (like 0001_waitlist); SQL content unchanged.

Group components/passkey/upsell.tsx + lib/passkey/aaguids.ts (were flat passkey-* files).
