+++ packages/db/drizzle/meta/_journal.json
@@ -20,7 +20,7 @@
       "idx": 2,
       "version": "7",
       "when": 1782569899122,
-      "tag": "0002_material_scorpion",
+      "tag": "0002_passkey",
       "breakpoints": true
     }
   ]
