+++ web/next/src/app/(protected)/dashboard/page.tsx
@@ -1,4 +1,4 @@
-import { PasskeyUpsell } from "@/components/passkey-upsell"
+import { PasskeyUpsell } from "@/components/passkey/upsell"
 
 export default function Page() {
   return (
